+++ app/javascript/src/navigation/components/Container.js
@@ -3,7 +3,7 @@ import 'shared/styles/globals'
 import React, { Component } from 'react'
 import { withRouter } from 'react-router-dom'
 import type { ContextRouter } from 'react-router-dom'
-import { Header } from './Header2'
+import { Header } from './Header'
 import { NotificationView, StickyContainer } from 'shared/components'
 import { stylesheet, mixins } from 'shared/styles'
 import { local } from 'shared/storage'
